requirements.txt
nltk>=3.8
sentence-transformers>=2.2.2
numpy>=1.24
pathlib>=1.0
python-dotenv>=1.0
torch>=2.0
transformers>=4.30
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=0.8.11
Pillow>=9.5.0
opencv-python>=4.8.0
spacy>=3.5.0